apply email template for buy package, waitlist and free class

+++ app/controllers/home.py
@@ -1,7 +1,7 @@
 from passlib.hash import bcrypt
 from app.models.users import User
 from app.models.packages import Package, UserPackage
-from app.helper import send_email_verification
+from app.helper import send_email_verification, send_email
 from app.models.schedules import InstructorSchedule, BookedSchedule
 from app.models.admins import Instructor, Admin
 from datetime import datetime
@@ -107,6 +107,10 @@ def buy(self):
                         transaction = yield transaction.save()
                         user = yield user.save()
 
+                        user = (yield User.objects.get(user._id)).serialize()
+                        content = str(self.render_string('emails/buy', user=user, host=self.request.host), 'UTF-8')
+                        yield self.io.async_task(send_email, user=user, content=content, subject='Bought Package')
+
                         self.redirect('/#/account#packages')
                     else:
                         self.set_status(403)
@@ -123,15 +127,20 @@ def test_waitlist(self):
 
     if self.get_argument('date'):
         date = datetime.strptime(self.get_argument('date'),'%Y-%m-%d');
-        sched_id = ObjectId(self.get_argument('sched_id'))
-        sched = yield InstructorSchedule.objects.get(sched_id)
+        sched = yield InstructorSchedule.objects.get(date=date)
         if not sched:
-            ins = yield Instructor.objects.limit(1).find_all()
-            if ins:
-                sched = InstructorSchedule(instructor=ins[0], type='regular', day='mon',
-                                           start=datetime.strptime('7:00 AM','%I:%M %p'), 
-                                           end=datetime.strptime('9:30 AM','%I:%M %p'))
-                sched = yield sched.save()
+            sched_id = ObjectId(self.get_argument('sched_id'))
+            if sched_id:
+                ins = yield Instructor.objects.limit(1).find_all()
+                if ins:
+                    sched = InstructorSchedule(instructor=ins[0], type='regular', day='mon',
+                                               start=datetime.strptime('7:00 AM','%I:%M %p'), 
+                                               end=datetime.strptime('9:30 AM','%I:%M %p'))
+                    sched = yield sched.save()
+            else:
+                self.write('Please provide a date')
+                self.finish()
+                return
 
         user = yield User.objects.get(email='user@waitlist.com')
         if not user:
@@ -153,8 +162,10 @@ def test_waitlist(self):
 
 def remove_test_waitlist(self):
     user = yield User.objects.get(email='user@waitlist.com')
-    yield BookedSchedule.objects.filter(user_id=user._id).delete();
-    yield user.delete()
+    if user:
+        yield BookedSchedule.objects.filter(user_id=user._id).delete();
+        yield user.delete()
+    self.finish()
 
 
 def add_regular_schedule(self):
